[#1414] moved apk update service to work manager
