+++ app/src/main/AndroidManifest.xml
@@ -98,14 +98,6 @@
         <service android:name=".service.BootstrapService" />
         <service android:name=".service.TimeCheckService" />
         <service
-            android:name=".service.ApkUpdateService"
-            android:exported="true"
-            android:permission="com.google.android.gms.permission.BIND_NETWORK_TASK_SERVICE">
-            <intent-filter>
-                <action android:name="com.google.android.gms.gcm.ACTION_TASK_READY" />
-            </intent-filter>
-        </service>
-        <service
             android:name=".service.FileChangeTrackingService"
             android:exported="true"
             android:permission="com.google.android.gms.permission.BIND_NETWORK_TASK_SERVICE">
